Move `ConnectionPdo` and `StatementPdo` to `Adapter/Pdo`.
These two classes were in the `Runtime\Connection` namespace, yet they are PDO implementations of
an interface defined in the connection namespace. When thinking about other adapters (like oci8),
it appears that these would need to have their own connection and statement implementations, and
therefore it makes sense to move the PDO ones into the `Propel\Runtime\Adapter\Pdo` namespace.
