+++ src/Propel/Runtime/Adapter/Pdo/PdoAdapter.php
@@ -12,7 +12,6 @@ namespace Propel\Runtime\Adapter\Pdo;
 
 use Propel\Runtime\Adapter\AdapterInterface;
 use Propel\Runtime\Adapter\AdapterException;
-use Propel\Runtime\Connection\ConnectionPdo;
 use Propel\Runtime\Connection\ConnectionInterface;
 use Propel\Runtime\Connection\StatementInterface;
 use Propel\Runtime\Map\ColumnMap;
@@ -35,7 +34,7 @@ abstract class PdoAdapter
      *
      * @param array    $conparams connection parameters
      *
-     * @return Propel\Runtime\Connection\ConnectionPdo
+     * @return Propel\Runtime\Adapter\Pdo\PdoConnection
      */
     public function getConnection($conparams)
     {
@@ -66,7 +65,7 @@ abstract class PdoAdapter
         }
 
         try {
-            $con = new ConnectionPdo($dsn, $user, $password, $driver_options);
+            $con = new PdoConnection($dsn, $user, $password, $driver_options);
             $this->initConnection($con, isset($conparams['settings']) && is_array($conparams['settings']) ? $conparams['settings'] : array());
         } catch (PDOException $e) {
             throw new AdapterException("Unable to open PDO connection", $e);
