+++ tests/Propel/Tests/Runtime/Connection/ConnectionManagerMasterSlaveTest.php
@@ -13,7 +13,6 @@ namespace Propel\Tests\Runtime\Connection;
 use Propel\Tests\Helpers\BaseTestCase;
 
 use Propel\Runtime\Connection\ConnectionManagerMasterSlave;
-use Propel\Runtime\Connection\ConnectionPdo;
 use Propel\Runtime\Adapter\Pdo\SqliteAdapter;
 
 use \PDO;
@@ -36,7 +35,7 @@ class ConnectionManagerMasterSlaveTest extends BaseTestCase
         $con = $manager->getWriteConnection(new SqliteAdapter());
         $this->assertInstanceOf('Propel\Runtime\Connection\ConnectionWrapper', $con);
         $pdo = $con->getWrappedConnection();
-        $this->assertInstanceOf('Propel\Runtime\Connection\ConnectionPdo', $pdo);
+        $this->assertInstanceOf('Propel\Runtime\Adapter\Pdo\PdoConnection', $pdo);
     }
 
     public function testGetWriteConnectionBuildsConnectionNotBasedOnReadConfiguration()
@@ -56,7 +55,7 @@ class ConnectionManagerMasterSlaveTest extends BaseTestCase
         $con = $manager->getReadConnection(new SqliteAdapter());
         $this->assertInstanceOf('Propel\Runtime\Connection\ConnectionWrapper', $con);
         $pdo = $con->getWrappedConnection();
-        $this->assertInstanceOf('Propel\Runtime\Connection\ConnectionPdo', $pdo);
+        $this->assertInstanceOf('Propel\Runtime\Adapter\Pdo\PdoConnection', $pdo);
     }
 
     public function testGetReadConnectionBuildsConnectionNotBasedOnWriteConfiguration()
