+++ tests/Propel/Tests/Runtime/Connection/ConnectionManagerSingleTest.php
@@ -13,7 +13,7 @@ namespace Propel\Tests\Runtime\Connection;
 use Propel\Tests\Helpers\BaseTestCase;
 
 use Propel\Runtime\Connection\ConnectionManagerSingle;
-use Propel\Runtime\Connection\ConnectionPdo;
+use Propel\Runtime\Adapter\Pdo\PdoConnection;
 use Propel\Runtime\Adapter\Pdo\SqliteAdapter;
 
 class ConnectionManagerSingleTest extends BaseTestCase
@@ -34,7 +34,7 @@ class ConnectionManagerSingleTest extends BaseTestCase
         $con = $manager->getWriteConnection(new SqliteAdapter());
         $this->assertInstanceOf('Propel\Runtime\Connection\ConnectionWrapper', $con);
         $pdo = $con->getWrappedConnection();
-        $this->assertInstanceOf('Propel\Runtime\Connection\ConnectionPdo', $pdo);
+        $this->assertInstanceOf('Propel\Runtime\Adapter\Pdo\PdoConnection', $pdo);
     }
 
     public function testGetReadConnectionReturnsWriteConnection()
@@ -48,7 +48,7 @@ class ConnectionManagerSingleTest extends BaseTestCase
 
     public function testSetConnection()
     {
-        $connection = new ConnectionPdo('sqlite::memory:');
+        $connection = new PdoConnection('sqlite::memory:');
         $manager = new ConnectionManagerSingle();
         $manager->setConnection($connection);
         $conn = $manager->getWriteConnection();
